minor fixes refactor
