+++ src/shared-components/index.js
@@ -12,7 +12,7 @@ export { default as Checkbox } from './checkbox';
 export { default as Chip } from './chip';
 export { default as Container } from './container';
 export { default as Dropdown } from './dropdown';
-export { default as Field } from './forms';
+export { default as Field } from './field';
 export { default as ImmersiveModal } from './immersiveModal';
 export { default as LoadingSpinner } from './loadingSpinner';
 export { default as Modal } from './modal';
